Get back to upstream
This change set aims get close to upstream again, before completely
diverging from the fork (see FUTUREWORK). This would ease a possible
rejoin in the future.

As a result, various changes revert recent work done in this fork. The
reasoning, in part, is not known anymore.

It also introduces some breaking changes. Namely, the interface for
multiple layouts. The layout name is derived from the respective key
in 'minio_layouts'. Additionally, the 'layouts' variable has been
renamed to follow a common role pattern, that is to prefix each one.

Instead of using the default test scenario, this introduces a dedicated
scenario to test the multi-layout setup. It is very similar to the
'default' scenario, but applies the role twice just with different layouts
and on less instances.

Please note, that atm the role only supports Ansible version >= 2.7, < 2.8

FUTUREWORK:

* upgrade to python3, ansible 2.10
* testing
  * move von travis to github actions
  * upgrade to Molecule 3
* refactor the the layouts idea, so that there is a default value for
  'minio_layout' instead of joining global attributes (e.g. 'minio_server_datadirs')
  with the corresponding ones defined for a specific layout